### 2019-10-25 #9 @mupsh
[Re] Mathematical analysis of depolarization block mediated by slow inactivation of fast sodium channels in midbrain dopamine neurons.
**Original article**: https://www.ncbi.nlm.nih.gov/pubmed/25185810

**Metadata URL**: https://github.com/mupsh/ReScience_Qian_2014/blob/master/article/metadata.yaml
**Code URL**: https://github.com/mupsh/ReScience_Qian_2014

**Scientific domain:** Computational Neuroscience
**Suggested editor:** Benoit Girard, Tiziano Zito
