### 2019-10-29 #9 @cJarvers
I'd be happy to review this as well, but I'll be very busy throughout November. Is early December an acceptable time frame for the review? Otherwise I have to pass.
