### 2019-10-29 #9 @rougier
Thanks @rgutzen and @cJarvers. @cJarvers I think we can wait  for you review in December yes.
@mupsh Is that ok for you ?
