### 2019-10-29 #9 @mupsh
Hi,

Yes, that’s ok for me.

Cheers,

Mary

> On Oct 29, 2019, at 11:07 AM, Nicolas P. Rougier <notifications@github.com> wrote:
>
> Thanks @rgutzen <https://github.com/rgutzen> and @cJarvers <https://github.com/cJarvers>. @cJarvers <https://github.com/cJarvers> I think we can wait for you review in December yes.
> @mupsh <https://github.com/mupsh> Is that ok for you ?
>
