### 2019-11-05 #9 @rgutzen
Hi @mupsh, I think you chose a nice study to replicate and I really hope I can help improve it towards a level of rigor, I think it could have. So, here my review.
You show some qualitative and some quantitative replications, but you also show some clear discrepancies that you either ignore or don't discuss sufficiently. I think there are a couple of major and minor changes that can make your replication study much stronger and more interesting and useful for the reader. In the following, I ordered my comments into general remarks, comments related to the code, comments regarding the replication of each figure, and comments related to the actual article text.

## General comments
* First of all, there is only a brief introduction, which doesn't answer the interesting question of what your motivation is for replicating exactly this study? And why is this a relevant paper? What are the key insights? And why would it be beneficial for the scientific community to replicate these results?
* There is no information on the reproducibility of the original implementation. Is the original data available? Is the original code available? Only at the very end, you mention the communication with the authors. This is very relevant to follow your replication and should be moved to the beginning of the paper and elaborated on.
* Ideally, you can also ask the original authors, if they agree to publish the code they provided you alongside this replication study
* Your paper is rather short, and writing short precise papers is generally desirable. However, you mostly just state what you replicated, and I'm totally missing a description of your journey how you replicated the results. What was the approach? What are the design principles or your implementation? Which difficulties did you encounter? What did not work? These aspects, I would consider the most valuable part of your research, and the ones others can really benefit and learn from.
* Especially, when you are not able to replicate a result or figure (exactly, or at all), it is not sufficient to say that you tried, it didn't work out, and then speculate that it might be due to different parameters. There needs to be a more detailed account of what exactly you tried, why you think the replication fails, and generally, you need to present a reasonable effort in investigating and fixing the failed replication.
* For each figure I'm missing an explanation of what exactly is shown and why this is interesting. This doesn't need to be long, but it is required in order to make sense of the replication, as this is the logical basis on which anyone can then start evaluate the similarity to the original results.
* For generating the simulation results, additional simulation parameters are set. Instead of hiding them in the code, they should be reported, and if necessary, discussed (e.g time resolution, input, discarded initialization time, ...)

## Code
### General
* The main goal of replication studies is to validate results from other publications by reimplementation. However, another goal is also to provide understandable and easily reproducible code so that others get better insights into the computational aspects of the original study and reuse/edit/expand the analysis or part of it.
The second goal requires the code to be in a certain format. If possible, it should use standards and existing software and packages which are common in the field. Furthermore, it needs to be logically structured and well documented.
Writing custom code to replicate other custom code misses the opportunity of adding real value to the implementation.
* So I strongly encourage you to use standard resources for the neuron models, like for example PyNN or NeuroML
* Besides documenting the code and structuring it in a way that makes it easy to read and understand, please also explain your code in the article. This is a computational paper and addressing and discussing certain implementation decisions in the text can be very helpful. The main outline of the implementation and the design principles could even be presented in their own Section.
* For the packages you do use, make sure you also mention them in the article, state what they do and why you chose them, especially for less common packages.
### Details
* executing run.py in the provided environment return an error while producing figure 3:
`unable to execute 'swig': No such file or directory
error: command 'swig' failed with exit status 1`
After installing swig version 3.0.12 it runs through.
* For understandability, it helps to assign variables with telling names instead of using numerical values in equations without explanations.
* code lines like `end_times = np.ravel(list(zip(np.arange(3, last_set + 100 + 1, 100), np.arange(100, last_set + 100 + 4, 100))))`
* for some reason you named tau_hs different from the other variables  as 'tah_hs'
* I strongly suggest making use of packages to handle units (like _quantities_ or _pints_) at least when variables are passed to functions. This easily ensures that the right magnitudes and units are used. Once the units are check in a function, they can then be omitted for the actual computations to not slow them down. Otherwise, it can get quite bothersome to trace back which magnitude the resulting values have.
* For a better overview and less error-proneness, it is generally advisable to outsource parameter settings (into a yamls, json, or event txt file), and import them and not set them by hand in each function.
* Also the `parameters` list is rather fragile since it relies on an arbitrary but specific order of the parameters. A dictionary would be a better choice.

## Replicating the figures
### General
* Using additional ticks (and tick labels) on the axis and would make the evaluation of the figures the comparison much easier.
* Use smaller linewidth, to not cover up the smaller details.
* Explain all the elements in each figure!
* Emphasize the relevant points in the plots. Discuss which aspects of the plots are the relevant results you aim to replicate?

### Figure 1a
* A legend is missing which of the curves is with and without hs.
* The code contains the following comment. `# need to divide the given value by 1.5 to get the correct graph`. This operation is not motivated or even mentioned in the text.
* The vertical axis misses its units
* I am missing a statement why you consider the plot replicated, like giving the agreement of characteristic points (e.g. minimum, horizontal trend in the beginning, Voltage at current drop-off, skewed parabola shape around minimum)

### Figure 1b
* even when the plot is mirrored in the x-direction, there are still clear differences (absolute and relative peak sizes, the distance between peaks) in the plot which needs to be addressed!
* Without a reason why the simulation would be backward in time, and given the other differences in the plot, the resulting plot suggests that there are other discrepancies in the equations or the code which need be uncovered. In this state, I would consider the plot not replicated.
* The code contains comment _# need to divide given value by 2 to get correct graph_. However, this is not mentioned in the text.
* g_Na is set to 5.92 nS, although the original article reports using 9.12 nS.
* The corresponding code is quite messy and thus hard to debug. To illustrate, I try to understand how the current trace which is plotted evolves. You are plotting the `clamp_current` which is calculated by `sodium_current()`, which takes `v`,`h`,and `hs` as state/time-dependent arguments. These are calculated by solving the `voltage_clamp` ODE using `ic` initial condition in the time packet `time`. Looking into `voltage_clamp()`, `__clamp__()` is called which calls the ODE function from the parameters list (here `ode_3d()`). `ode_3d()` itself is not easy to understand and debug since it has several not intuitive equations which stretch over multiple line. But most remarkably `ode_3d()` does not use its time argument `t`, which essentially means that the whole current function does not depend on the time packet it is iterated over.
The main point here is that you organize and structure the code obfuscates dependencies and makes it really hard to understand what is happening in the code.

### Figure 1c
* it should be made clear that the time course of n(h), which is not a function, is fitted by a function f(h)
* "to get the same results" Figure 1C is visibly different to the original!!
* The n-h relation in the simulation varies over time. See the spread of the curve in the original plot. This is not visible in the replication and not commented.
* The characteristic point where the n-h curve crosses itself is shifted significantly from around 0.2 to around 0.5 (also due to line width not well identifiable)
* It is totally unclear how you motivate the change in the equation for tau_n.
* This clear discrepancy of the n-h curve to the original and the replotting of a fitted function f(h) unrelated to the shown n-h curve does satisfy any criteria of a replication.
* If these results are indeed not replicable the authors should state so. However, they should put a reasonable effort into an adequate replication (e.g. by motivated changes in the equations/implementation) and document this process in detail.
* For the trace of f(h) in Figure 1C, you simply re-plot a polynomial with the reported fit parameters. However, these fit results are a result of the paper and should themselves be replicated.

### Figure 1D1
* The time axis is missing
* Missing comparison of essential features and why you consider the plot replicated.
* Please comment on the initial dip of the trace which is lower than the others.

### Figure 1D2
* The time axis is missing
* Missing comparison of essential features and why you consider the plot replicated.
* Please comment on the increasing peak and decreasing trough values. This trend is also influenced by the change you applied to the time constant function for _n_, please consider this as well.

### Figure 2
* missing time axis
* discrepancy in the scale of the voltage, please mention and discuss
* please discuss the shorter 'ringing' time in you simulation

### Figure 3
* For a full replication, the figure is missing the part showing the instantaneous firing rate after the current step.
* For figures 3B the relation between figures 3A and the choice of `hs` values does not become clear. Please, motivate this in the caption.
* Figure 3C
    * This figure is missing an explanation of its elements, especially the newly introduced ones (either as legend or in the caption): solid line, dashed line, pink line, H2, red points.
    * Why did you choose to extend the hs axis with respect to the original?
    * the grey trajectory line covers the most crucial point of the plot!
    * The linewidth of the trajectory and the contraction of the region of interest due to the larger scale make the trajectory dynamics unnecessary hard to see.
* For handling the ODEs you use scipy functions and PyDSTool. Please, state this explicitly in the text and explain what exactly you use them for and why you chose them.

### Figure 4
* missing h axis for Figure 4 A2
* missing arrow or other indication in which direction the current is increased. Just saying from left to right or right to left is not very understandable on its own, especially since you didn't put in the fixpoints.
* Figure B: missing explanation of the elements!
* In Figure B2 again the region around the subcritical Hopf point is covered by large red markers and overlapping, unreadable labels.
* Missing current units for Figure B
* For Figure 4C, please state which current and which voltage you plot, why this is of interest.
* Missing current axis for Figure 4 C2

### Figure 5
* missing value of depolarization in all plots
* Again, there is missing an explanation of the key results which are represented in these plots, which you are trying to replicate. Without this description, a reader can a) hardly understand the figure and b) can not reasonably assess if and why the replication failed.
* Your code also doesn't allow for easy bug tracking here, due to not understandable or even misleading variable names (the solution should be a voltage but is named `ic`), missing clarifying comments, and convoluted structure.

## Article text
* Why do you cite the original paper as 'Levitan et al. (2014)' although the first author is Qian?
* Commonly in articles, you present you the work in the present tense. Here, it can make sense to refer to the original work in the past tense, but at for least your results and figures it would make sense to use the present tense. Please review, and use it consistently.
* You use 'replicated' and 'reproduced' interchangeably throughout the text. Please adhere to the definitions of ReScience use the correct terminology. Reproduction: Running the same computation on the same data; Replication: repeating a computational study in the spirit of a published protocol, but with a different realization. (https://rescience.github.io/faq/)
* You use a lot of emphasizing adjectives like 'successfully' or 'faithfully'. Instead of using these formulations, describe the replication in a more neutral and precise manner. State exactly the argumentation why and to which degree you consider a result as replicated.
* reference the location of the equation you modified in the original paper
* There is a wrong minus sign in equation (compare to the original)! It is however correct in the code.
* explain the variables in the equation
* You need to explain you motivation why you modified this time constant equation and you need to argue why exactly you did change this numeric value from 40 to 60, instead of many other changes you could have made, and why the modified version should be considered 'better' than the original.
* Figure 1 should be moved into Results section, after the first paragraph
* "All original figures, [..], were successfully replicated". Not true. At least the last figure was not replicated!
* "..., successfully showing that they had not compromised the original sodium current description." Elaborate. What constitutes 'not compromised' here?
* "seemingly identical results" Be precise how they differ, how you compared the results, and on what basis (according to which features) you consider them equal.
* 'the same but “backwards” or ”flipped”' Be more precise. Along which axis? According to which attribute you consider it being 'flipped'?
* In many sentences there are missing articles, for example:
    * ".. when depolarizing current was repeatedly applied"
    * ".. modified potassium current"
    * ".. plotted membrane current"
* "in our implementation". Either 'with' our implementation, or in our simulation.
* In the discussion section, you mention that the original code contains different parameters than the paper. Please, state also these parameters and include this in the discussion about which equation to choose, and why and how to modify it.
* Use consistent Figure names. Either 'Figure 2 B1' or 'Figure 2B1'.
* "More specifically, our replication of Figure 1B was “backwards” or
”flipped” where in our replication the sodium current amplitude in time, which is opposite in the original implementation where it decreases towards 0 in time." This sentenced is very unclear, please re-formulate.
* "Hence, we can only speculate that this may have potentially been a graphing error." There are clear qualitative differences in the plots (see above) which can not be explained by 'graphing error', please elaborate.
* "Overall, this implementation replicated the original paper."/"With the exception of Figure 6, this implementation confidently shows that the original results were correctly implemented" No, you present a partial replication and found several discrepancies. Be more precise.
