### 2019-11-06 #9 @rougier
Thanks @rgutzen for this very detailed review. @mupsh You can wait for the second review or you can start modifying your article, taking @rgutzen comments and suggestions into account.
