### 2019-11-22 #9 @mupsh
@rgutzen Thank you for your extensive comments and we're working on changes. @cJarvers Since we're going to be making fairly extensive changes it might be easiest for us to add our changes so you don't have to go through 2 sets.
