### 2019-12-19 #9 @rougier
@mupsh Did you have time to update the manuscript?
