### 2019-12-20 #9 @cJarvers
Yes, first week of January is fine with me.
