### 2020-01-23 #9 @cJarvers
Just to make sure we're all on the same page: @rougier, would you like the review by me to be done on Monday or the revisions by @mupsh? I have not done the review yet since the authors suggested I wait for their revisions first. If that is a problem, I can try to do the review by Monday.
