### 2020-01-27 #9 @maryupshall
Sorry for the delay in getting this out, and thanks to @rgutzen for his thorough review. Overall we've implemented most of the requested changes and believe this puts the manuscript in a better state. For clarity we've included the original comments along with our reply. @cJarvers we're ready to move on with the review now.

# General Comments

- First of all, there is only a brief introduction, which doesn't answer the interesting question of what your motivation is for replicating exactly this study? And why is this a relevant paper? What are the key insights? And why would it be beneficial for the scientific community to replicate these results?
    - We’ve expanded the introduction to give context for why these neurons and depolarization block are of interest with regards to schizophrenia and the authors conclusions.

- There is no information on the reproducibility of the original implementation. Is the original data available? Is the original code available? Only at the very end, you mention the communication with the authors. This is very relevant to follow your replication and should be moved to the beginning of the paper and elaborated on.
    - We mention that they’ve provided us with the model implementation they had available.

- Ideally, you can also ask the original authors, if they agree to publish the code they provided you alongside this replication study.
    - We don’t believe this would be of interest as the code provided is only the implementation (no analysis of manipulations) of the 3d model

- Your paper is rather short, and writing short precise papers is generally desirable. However, you mostly just state what you replicated, and I'm totally missing a description of your journey how you replicated the results. What was the approach? What are the design principles or your implementation? Which difficulties did you encounter? What did not work? These aspects, I would consider the most valuable part of your research, and the ones others can really benefit and learn from.
    - We appreciate that our original manuscript was too sparse, we’ve elaborated on the replication in the modified version. We are unclear as to what you mean by design principles or approach, this could mean many things most of which we believe are outside the scope of a replication paper.

- Especially, when you are not able to replicate a result or figure (exactly, or at all), it is not sufficient to say that you tried, it didn't work out, and then speculate that it might be due to different parameters. There needs to be a more detailed account of what exactly you tried, why you think the replication fails, and generally, you need to present a reasonable effort in investigating and fixing the failed replication.
    - We provided more detail, and a justification of why the original parameters are incorrect.

- For each figure I'm missing an explanation of what exactly is shown and why this is interesting. This doesn't need to be long, but it is required in order to make sense of the replication, as this is the logical basis on which anyone can then start evaluate the similarity to the original results.
    - Added explanation of what is shown.

- For generating the simulation results, additional simulation parameters are set. Instead of hiding them in the code, they should be reported, and if necessary, discussed (e.g time resolution, input, discarded initialization time, ...)
    - We’ve elaborated on our parameters in the methods section.

# General

- The main goal of replication studies is to validate results from other publications by reimplementation. However, another goal is also to provide understandable and easily reproducible code so that others get better insights into the computational aspects of the original study and reuse/edit/expand the analysis or part of it.
The second goal requires the code to be in a certain format. If possible, it should use standards and existing software and packages which are common in the field. Furthermore, it needs to be logically structured and well documented.
Writing custom code to replicate other custom code misses the opportunity of adding real value to the implementation.
- So I strongly encourage you to use standard resources for the neuron models, like for example PyNN or NeuroML
    - While in general we agree with this statement, in this case we have to disagree. Due to the issues we’ve faced in replicating the manuscript, not having duplicate representations of the model (i.e. for time-domain solutions, and bifurcation analysis) is critical because if one implementation works but the other does not that adds another level of potential implementation error.
    - Since (to our knowledge) no neuron specification format can be parsed by any numerical bifurcation engine. Our choices are either implement the model in python and inject symbolic variables to prevent duplication which we’ve done. Or, write our own parser which would add far more problems than it would solve. For this reason, we’ve chosen to keep our implementation as is.

- Besides documenting the code and structuring it in a way that makes it easy to read and understand, please also explain your code in the article. This is a computational paper and addressing and discussing certain implementation decisions in the text can be very helpful. The main outline of the implementation and the design principles could even be presented in their own Section.
    - We agree entirely that our initial submission should not have been presented in the state that it was in. We’ve now thoroughly refactored it and documented it properly. That being said, we’re unsure what you mean by discussing design principles in the paper. Since the goal of the work is, in-essence, to validate the original work. We’re unclear as to the benefit of discussing for example OOP vs imperative implementations. Could you please provide an example of what you would like?

- For the packages you do use, make sure you also mention them in the article, state what they do and why you chose them, especially for less common packages.
    - We’ve implemented this in the methods.

# Details

- executing run.py in the provided environment return an error while producing figure 3:
unable to execute 'swig': No such file or directory error: command 'swig' failed with exit status 1
After installing swig version 3.0.12 it runs through.
    - We didn’t catch this omission and it’s now specified in our readme.

- For understandability, it helps to assign variables with telling names instead of using numerical values in equations without explanations.
- code lines like end_times = np.ravel(list(zip(np.arange(3, last_set + 100 + 1, 100), np.arange(100, last_set + 100 + 4, 100))))
    - Very true, this again is an oversight on our behalf, and we’ve refactored out these obscure definitions.

- for some reason you named tau_hs different from the other variables as 'tah_hs'
    - This was a typo and it’s been corrected.

- I strongly suggest making use of packages to handle units (like quantities or pints) at least when variables are passed to functions. This easily ensures that the right magnitudes and units are used. Once the units are check in a function, they can then be omitted for the actual computations to not slow them down. Otherwise, it can get quite bothersome to trace back which magnitude the resulting values have.
    - In general, yes this is true. However, in this case since the parameter scheme is consistent, except when the authors give “somewhat obviously incorrect units” (which a computational unit system would not help) such as nS/cm^2 we believe won’t really help in this case.

- For a better overview and less error-proneness, it is generally advisable to outsource parameter settings (into a yamls, json, or event txt file), and import them and not set them by hand in each function.
    - Since this model implementation is not dynamic i.e. it’s not changing. Many benefits of an external config are lost. There are two options in this case for implementing config files, either duplicating all parameters in each model’s config which we don’t want to do (see neuroML comment for duplication argument) or each specification would only be used for setting non-defaults. In which case you’d still have to check back to either the default function or the default config file which in our view is just as cumbersome as looking up the function call for non-default arguments. For this reason, we will leave our implementation as is

- Also the parameters list is rather fragile since it relies on an arbitrary but specific order of the parameters. A dictionary would be a better choice.
    - This is completely true and we’ve implemented this.

# General
- Using additional ticks (and tick labels) on the axis and would make the evaluation of the figures the comparison much easier.
- Use smaller linewidth, to not cover up the smaller details.
- Explain all the elements in each figure!
- Emphasize the relevant points in the plots. Discuss which aspects of the plots are the relevant results you aim to replicate?
    - We’ve cleaned up our plots where possible and discussed them in the manuscript. We also made sure to explain all the elements of each figure in the figure captions or using a legend.

# Figure 1a

- A legend is missing which of the curves is with and without hs.
- The code contains the following comment. # need to divide the given value by 1.5 to get the correct graph. This operation is not motivated or even mentioned in the text.
- The vertical axis misses its units
- I am missing a statement why you consider the plot replicated, like giving the agreement of characteristic points (e.g. minimum, horizontal trend in the beginning, Voltage at current drop-off, skewed parabola shape around minimum)
    - We’ve implemented these changes as requested.

# Figure 1b

- even when the plot is mirrored in the x-direction, there are still clear differences (absolute and relative peak sizes, the distance between peaks) in the plot which needs to be addressed!
- Without a reason why the simulation would be backward in time, and given the other differences in the plot, the resulting plot suggests that there are other discrepancies in the equations or the code which need be uncovered. In this state, I would consider the plot not replicated.
- The code contains comment # need to divide given value by 2 to get correct graph. However, this is not mentioned in the text.
- The corresponding code is quite messy and thus hard to debug. To illustrate, I try to understand how the current trace which is plotted evolves. You are plotting the clamp_current which is calculated by sodium_current(), which takes v,h,and hs as state/time-dependent arguments. These are calculated by solving the voltage_clamp ODE using ic initial condition in the time packet time. Looking into voltage_clamp(), __clamp__() is called which calls the ODE function from the parameters list (here ode_3d()). ode_3d() itself is not easy to understand and debug since it has several not intuitive equations which stretch over multiple line. But most remarkably ode_3d() does not use its time argument t, which essentially means that the whole current function does not depend on the time packet it is iterated over.
The main point here is that you organize and structure the code obfuscates dependencies and makes it really hard to understand what is happening in the code.
    - With regards to the “mirroring” upon re-reading their definition and properly re-implementing the code we’ve fixed this issue.
    - As discussed in the text there is a scaling issue which we’ve addressed as well as we could by manipulating the experimental parameters
    - Your comments about the code quality and smells are correct and we’ve refactored these. To clarify what’s happening in the call stack is that: clamp (ignore the dunder) is generic function for clamping a variable and voltage_clamp is the specific wrapper for a voltage clamp (instead of clamping another variable). The comment about not using time is subtle we did not do a good job explaining it in the code in the initial version. Since the pulses of current or clamping are discrete, they shouldn’t be implemented as a non-autonomous function, therefore it’s implemented as a sequence of solved autonomous ODEs with the initial condition of the subsequent step being the end of the current step, hence why “t” is unused. We’ve documented and refactored this in a way that we believe makes this clearer.

# Figure 1c

- it should be made clear that the time course of n(h), which is not a function, is fitted by a function f(h)
- "to get the same results" Figure 1C is visibly different to the original!!
- The n-h relation in the simulation varies over time. See the spread of the curve in the original plot. This is not visible in the replication and not commented.
- The characteristic point where the n-h curve crosses itself is shifted significantly from around 0.2 to around 0.5 (also due to line width not well identifiable)
- It is totally unclear how you motivate the change in the equation for tau_n.
- This clear discrepancy of the n-h curve to the original and the replotting of a fitted function f(h) unrelated to the shown n-h curve does satisfy any criteria of a replication.
- If these results are indeed not replicable the authors should state so. However, they should put a reasonable effort into an adequate replication (e.g. by motivated changes in the equations/implementation) and document this process in detail.
- For the trace of f(h) in Figure 1C, you simply re-plot a polynomial with the reported fit parameters. However, these fit results are a result of the paper and should themselves be replicated.
    - We’ve implemented the changes as requested however to address a few specific points. And we outright say that we could not match the authors’ implementation
    - The n-h limit cycle (which is a limit cycle) appears to change over time as they’ve included a transient. And we don’t discuss this as 1) that’s standard for displaying limit cycles and 2) there are far greater issues we wanted to focus on.
    - The main issue here is that the function tau_n cannot be the correct function due to an invalid domain (see discussion in the paper) we’ve motivated our modification and furthermore show in the supplemental the result with their original equation.
    - With regards to their polynomial, we do re-fit it (terms are printed during the run for figure 1) since we’ve established that the implementation cannot be correct, the n-h space space will be quite different as well (which it is), therefore any fit polynomial will be quite different. The important part is that in using their reduced polynomial model elsewhere we replicate those results.

# Figure 1D1

- The time axis is missing
- Missing comparison of essential features and why you consider the plot replicated.
- Please comment on the initial dip of the trace which is lower than the others.
    - Implemented as requested, the initial dip was a transient which has now been discarded as discussed in the MS.
    - Additionally, we had flipped the location of 1D1 and 1D2 this has been corrected.

# Figure 1D2

- The time axis is missing
- Missing comparison of essential features and why you consider the plot replicated.
- Please comment on the increasing peak and decreasing trough values. This trend is also influenced by the change you applied to the time constant function for n, please consider this as well.
    - Again, this trend was a transient which has been corrected.

# Figure 2

- missing time axis
- discrepancy in the scale of the voltage, please mention and discuss
- please discuss the shorter 'ringing' time in you simulation
    - Addressed & discussed.

# Figure 3

- For a full replication, the figure is missing the part showing the instantaneous firing rate after the current step.
- For figures 3B the relation between figures 3A and the choice of hs values does not become clear. Please, motivate this in the caption.
- Figure 3C
    - This figure is missing an explanation of its elements, especially the newly introduced ones (either as legend or in the caption): solid line, dashed line, pink line, H2, red points.
    - Why did you choose to extend the hs axis with respect to the original?
    - the grey trajectory line covers the most crucial point of the plot!
    - The linewidth of the trajectory and the contraction of the region of interest due to the larger scale make the trajectory dynamics unnecessary hard to see.
- For handling the ODEs you use scipy functions and PyDSTool. Please, state this explicitly in the text and explain what exactly you use them for and why you chose them.
    - We’ve added the figure components as discussed. We’ve also discussed their use. However, we don’t discuss why we’ve used them as really there is no other viable option. We are not aware of any other established ode solvers in python asides from scipy odeint. And there are a few XPPAUT / AUTO python interfaces however they’re all at least several years without maintenance and are often undocumented, hence PyDSTool is the only viable option. It seems slightly unnecessary to state we used “x” because there’s nothing else.
    - For the relation between figures 3A and the choice of hs values in 3B, these come from the “distinct” regimens in A2. We don’t address this as the goal is to show the change in stability over the depolarization block hence the parameters are somewhat arbitrary. Furthermore, this a scientific discussion which we feel is outside the domain of the replication.
    - We’ve addressed the other display points.
    - Added explanation of the elements in the figure caption.

# Figure 4

- missing h axis for Figure 4 A2
- missing arrow or other indication in which direction the current is increased. Just saying from left to right or right to left is not very understandable on its own, especially since you didn't put in the fixpoints.
- Figure B: missing explanation of the elements!
- In Figure B2 again the region around the subcritical Hopf point is covered by large red markers and overlapping, unreadable labels.
- Missing current units for Figure B
- For Figure 4C, please state which current and which voltage you plot, why this is of interest.
- Missing current axis for Figure 4 C2
    - Added explanation of the elements in the Figure caption.
    - Fixed and added explanation of the elements.
    - With regards to the bifurcation we are somewhat limited in terms of control of elements and features (such as plot order). That said we’ve done what we can to maximize readability.

# Figure 5

- missing value of depolarization in all plots
- Again, there is missing an explanation of the key results which are represented in these plots, which you are trying to replicate. Without this description, a reader can a) hardly understand the figure and b) can not reasonably assess if and why the replication failed.
- Your code also doesn't allow for easy bug tracking here, due to not understandable or even misleading variable names (the solution should be a voltage but is named ic), missing clarifying comments, and convoluted structure.
    - Again, we agree the code was confusing and should not have been shipped in that state. It’s been refactored. While no-one should have been expected to understand the code as is. The “ic” (lines 46 to 49) was again used as the initial condition for the next step (due to the discrete conductance jump) and the state (i.e. the current solution) was concatenated with the growing full solution. The new implementation makes this much clearer.

# Article text

- Why do you cite the original paper as 'Levitan et al. (2014)' although the first author is Qian?
    - This was a copying error; we’ve fixed the references. @rougier we can fix the repo name, is it O.K. to edit this submission to reflect this?

[…]
