### 2020-02-09 #9 @cJarvers
Okay, @mupsh , here's my review:

This review is relative to commit https://github.com/mupsh/ReScience_Levitan_2014/commit/d5c641ffd14d5740e032e2864083215ac49dec73

## Reproducibility

I ran the code using Python 3.6.10, swig 4.0.1, PyDSTool 0.90.3, numpy 1.18.1, scipy 1.3.0 and matplotlib 3.1.1.

The code produced a warning message ("Test function going crazy: class Hopf_Bor(BordermMethod, BiAltMethod)"), but otherwise ran without problems and produced the figures as shown in the paper.

## Replication

The core results of the original paper are replicated. This requires some major parameter changes, so I'm not sure whether I share your optimistic conclusion (that "the 2D and 3D models were implemented correctly in the original work"), but I find the explanations of the discrepancies and your overall reasoning clear enough to accept this as a replication - thanks to your work I'm sufficiently confident that the conclusions of the original article were correct, even if there were errors in the implementation or the reporting.

## Clarity of Article

In general, I find the article is now a good deal clearer than before the revision. The explanations of what each figure shows and what adjustments were necessary to replicate each are easy to follow.

Unfortunately, the article is not self-contained. In order to understand the explanations of the replication, it is necessary to know the original article, which is behind a paywall. I could not find any guidelines about how self-contained the article should be (do you know any, @rougier?) and I guess most readers interested in a replication would know the original. Still, I think it would be good have a section that explains the model itself. This would enable to reader to better understand what is going on, without going back to the original article.

I also concur with @rgutzen that adding a section that explains the code would be helpful. This should not include overly general things like OOP vs. imperative programming (and I don't think that was meant with "design principles") and should be more aimed towards giving the reader a starting point to exploring/understanding the code. This would essentially be a more elaborate version of the first paragraph of your methods section (explaining why and how you used which package), including some pointers like the fact that you factorized your code into the differential equations for all three models (`diff_eq.py`), the steady-state equations and time constants (`gating.py`), the membrane currents (`current.py`) and nullclines (`nullclines.py`).

Another thing that could improve the clarity of the article would be to list all the changes made to the original model and between different figure in one central place, e.g., a table. Unless I missed something, these changes are:

- the changed formula for tau_n (used everywhere)
- the value of g_Na in Figure 1A
- the value of g_Na and the pulse width in Figure 1B
- the newly fitted polynomial coefficients in Figure 1C; the fitted values should be reported in the text
- the reversal potentials for NMDA and AMPA receptors in Figure 6
- the conductances g_AMPA and g_NMDA in Figure 6A and 6B (different values)

Especially for the sodium conductance, it would be good to know what value you used for different simulations, the one from 1A or 1B.

Otherwise, I noticed a few smaller issues:

- Use of tenses is sometimes inconsistent. For example, in the second paragraph one sentence starts with "Previous work had", whereas the next starts with "Previous work has".
- The formulation "This new model included, or exclused a slow inactivation term" might confuse readers. It's not clear that this relates to the 2D and 3D versions mentioned before.
- The legend in Figure 1A covers a small part of the curve.
- The "ringing" in Figure 2 is hardly visible. It might help to make the figure bigger and possibly decrease the line width.
- In the text, the term "ringing" is not explained, except by referring to Figure 2. A brief definition might help.
- In the explanation for Figure 4, you mention that you use a different naming convention (SNP vs. LPC) but do not explain why.
- The last sentence of section 2.4 is incomplete.

## Clarity of Code

The code is well structured and commented. The separation into simulation code, runners for the individual figures and plot setup works well and helped me find my way around the code.

In two places, this wasn't quite true: The `diff_eq.py` module contains a lot of stuff, some of which is not directly related to the differential equations and could for example go to an `experiments.py` module (e.g. the clamp experiments or IV curves). The other place is the functions for figures 3c and 4b, due to the use of pyDSTool. I don't really know how to separate that out better, though.

I also agree with @rgutzen that using packages for handling / checking units would be useful. This is especially true since some of the differences from the original paper involve scaling, so it would be good to make sure this is not due to some implicit unit conversion error somewhere.

One minor point I noticed: the explanation of parameter `parameter_name` in `diff_eq.pulse` is only "Parameter to"; there's probably something missing there.

## Summary

Overall, I'm quite happy with the paper. I think some improvements are possible: (1) explaining the model in the text to make the article more self-contained, (2) listing changes and parameter values in one place / table and (3) adding explicit unit handling to the code.
