### 2020-02-28 #9 @rougier
thanks @cJarvers. @mupsh can you address the comments ?

@mupsh Don't forget to change the title!

@cJarvers I think you're right and we may need to add a section for aiming at self-contained article when the original article is not available (behind a paywall). Could you propose some changes authors instruction based on your experience ?

**Note**  I found this link which seems to be free: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4254877/ (or is it just me ?)
