### 2020-03-02 #9 @cJarvers
@rougier True, the ncbi-link is free. Question is whether green OA like this solves the issue, since readers still need to find the free version. Maybe adding a link to the replication article would work. Anyways, I'll open a separate issue to suggest some changes to the author instructions.
