### 2020-04-24 #9 @cJarvers
Thanks for your reply, @mupsh. I'm quite happy with how you addressed my comments.

I ran the code again (with slightly different versions of the dependencies: python 3.8.2, scipy 1.4.1, matplotlib 3.2.1, sympy 1.5.1, Pint 0.11, PyDSTool 0.91) and everything behaves as expected.

There are only two minor things I noticed in the article:

- In the very first line, the word "dopaminergic" stretches over the right border. Not sure how hard that is to fix.
- Reference 11 seems to be incomplete.

@rougier After these minor things are fixed, I'd recommend acceptance.
