### 2020-04-27 #9 @rougier
@cJarvers @rgutzen Thanks for your very nice and extensive reviews. I therefore accept the manuscript that will be published soon hopefully. Congratulations @mupsh 🎉 !

I will make a PR to your repository to complete your metadata. In the meantime, can you register your code repository at https://www.softwareheritage.org/save-and-reference-research-software/. You should obtain a swid that you can post here.
