### 2020-04-30 #9 @maryupshall
@rougier Great news, thank you very much! We've made the minor fixes, I've registered the code as requested and obtained the following SWH-ID:
https://archive.softwareheritage.org/swh:1:rev:e1040fe3fbaefef4166964772efa08ada444db6e;origin=https://github.com/mupsh/ReScience_Qian_2014.git/
