### 2020-05-03 #9 @rougier
I created a PR (https://github.com/mupsh/ReScience_Qian_2014/pull/10) for small changes to your repo before publication. Can you check it?
