### 2020-05-12 #9 @rougier
Ok, I just saw PR has been merged. I'll try to publish the paper by the end of the week. Don't hesitate to spam me if it's not done :)
