### 2020-05-25 #9 @rougier
It has been published and online ! Sorry for the delay.
See https://zenodo.org/record/3842360
